Added the ability to win by 5 in a row.

+++ src/org/us/_42/laphicet/gomoku/GameStateReporter.java
@@ -4,13 +4,13 @@ import java.util.Collection;
 
 public interface GameStateReporter {
 	/**
-	 * Handles a placed piece and other events that occurred on a given turn.
+	 * Reports a placed token and other events that occurred on a given turn.
 	 * 
 	 * @param game The Gomoku game controller including the game board.
-	 * @param x The x coordinate of the placed piece.
-	 * @param y The y coordinate of the placed piece.
-	 * @param value The value of the placed piece.
+	 * @param x The x coordinate of the placed token.
+	 * @param y The y coordinate of the placed token.
+	 * @param value The value of the placed token.
 	 * @param logs A collection of logs detailing events.
 	 */
-	public void logTurn(Gomoku game, int x, int y, byte value, Collection<String> logs);
+	public void logTurn(Gomoku game, int x, int y, int value, Collection<String> logs);
 }
